Move GraphQL `resolvers` to its own file
